Extract executable persistence helpers
